requirements.txt
aiohttp~=3.5.4
aiohttp-jinja2~=1.1.1
pyyaml~=5.1
